Split *.similarity.jamopp.parser package
