+++ commit-based-cipm/fi-tests/cipm.consistency.fitests.repositorytests/src/cipm/consistency/fitests/repositorytests/RepoTimeMeasurementTag.java
@@ -1,6 +1,6 @@
 package cipm.consistency.fitests.repositorytests;
 
-import cipm.consistency.fitests.similarity.jamopp.parser.ITimeMeasurementTag;
+import cipm.consistency.fitests.similarity.jamopp.parser.timemeasurement.ITimeMeasurementTag;
 
 /**
  * An enum containing various GIT-Repository-related tags that can be used while
