+++ commit-based-cipm/fi-tests/cipm.consistency.fitests/src/cipm/consistency/fitests/similarity/jamopp/parser/resultprovider/package-info.java
@@ -1,0 +1,1 @@
+package cipm.consistency.fitests.similarity.jamopp.parser.resultprovider;
